Renamed v1 modules, added PDF MIME code
